# pecus.Frontend/.spec/external-api-openapi-scheme.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Coati",
    "description": "AIを社畜扱いして作成するWebAPIです。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7265/"
    }
  ],
  "paths": {
...
    "/api/external/workspaces/{workspaceIdOrCode}/items": {
      "get": {
        "tags": [
          "ExternalAPI"
        ],
        "summary": "指定ワークスペース内の全アイテムを取得する",
        "description": "ページネーション（1ページあたり20件固定）に対応しています。\nアクティブフラグ（isActive）、アーカイブフラグ（isArchived）、下書きフラグ（isDraft）でのフィルタリングが可能です。\n認証された API キーの組織に属するワークスペースのみアクセス可能です。",
        "parameters": [
          {
            "name": "workspaceIdOrCode",
            "in": "path",
            "description": "ワークスペースIDまたはコード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "ページ番号（1始まり、省略時は1）",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "アクティブフラグフィルタ（nullの場合は全件）",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isArchived",
            "in": "query",
            "description": "アーカイブフラグフィルタ（nullの場合は全件）",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isDraft",
            "in": "query",
            "description": "下書きフラグフィルタ（nullの場合は全件）",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "アイテム一覧の取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalItemListResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ワークスペースが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/external/workspaces/{workspaceIdOrCode}/items/{itemNumber}": {
      "get": {
        "tags": [
          "ExternalAPI"
        ],
        "summary": "指定ワークスペース内の指定アイテムを取得する",
        "description": "ワークスペースIDまたはコードとアイテム番号（ワークスペース内連番）を指定してアイテム情報を取得します。\n本文は Markdown 形式に変換されて返却されます。\n認証された API キーの組織に属するワークスペース・アイテムのみアクセス可能です。",
        "parameters": [
          {
            "name": "workspaceIdOrCode",
            "in": "path",
            "description": "ワークスペースIDまたはコード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemNumber",
            "in": "path",
            "description": "アイテム番号（ワークスペース内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "アイテムの取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalItemResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ワークスペースまたはアイテムが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/external/workspaces/{workspaceIdOrCode}/items/{itemNumber}/tasks": {
      "get": {
        "tags": [
          "ExternalAPI"
        ],
        "summary": "指定ワークスペース・アイテム内のタスク一覧を取得する",
        "description": "指定されたワークスペースおよびアイテムに紐づくタスクの一覧を取得します。\n完了状態（isCompleted）や破棄状態（isDiscarded）によるフィルタリングが可能です。\n認証された API キーの組織に属するワークスペース・アイテムのみアクセス可能です。",
        "parameters": [
          {
            "name": "workspaceIdOrCode",
            "in": "path",
            "description": "ワークスペースIDまたはコード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemNumber",
            "in": "path",
            "description": "アイテム番号（ワークスペース内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "isCompleted",
            "in": "query",
            "description": "完了フラグフィルタ（nullの場合は全件）",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isDiscarded",
            "in": "query",
            "description": "破棄フラグフィルタ（nullの場合は全件）",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "タスク一覧の取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalWorkspaceTaskListResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ワークスペースまたはアイテムが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/external/workspaces/{workspaceIdOrCode}/items/{itemNumber}/tasks/{sequence}": {
      "get": {
        "tags": [
          "ExternalAPI"
        ],
        "summary": "指定ワークスペース・アイテム内の指定タスクを取得する",
        "description": "ワークスペースIDまたはコード、アイテム番号、タスクシーケンス番号を指定してタスク詳細を取得します。\n認証された API キーの組織に属するワークスペース・アイテム・タスクのみアクセス可能です。",
        "parameters": [
          {
            "name": "workspaceIdOrCode",
            "in": "path",
            "description": "ワークスペースIDまたはコード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemNumber",
            "in": "path",
            "description": "アイテム番号（ワークスペース内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sequence",
            "in": "path",
            "description": "タスクシーケンス番号（アイテム内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "タスクの取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalWorkspaceTaskDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ワークスペース、アイテム、またはタスクが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    },
    "/api/external/workspaces/{workspaceIdOrCode}/items/{itemNumber}/tasks/{sequence}/comments": {
      "get": {
        "tags": [
          "ExternalAPI"
        ],
        "summary": "指定ワークスペース・アイテム・タスク内のコメント一覧を取得する",
        "description": "指定されたタスクに紐づく全コメントを取得します（論理削除されたコメントを除く）。\n認証された API キーの組織に属するワークスペース・アイテム・タスクのみアクセス可能です。",
        "parameters": [
          {
            "name": "workspaceIdOrCode",
            "in": "path",
            "description": "ワークスペースIDまたはコード",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemNumber",
            "in": "path",
            "description": "アイテム番号（ワークスペース内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sequence",
            "in": "path",
            "description": "タスクシーケンス番号（アイテム内連番）",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "タスクコメント一覧の取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalTaskCommentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ワークスペース、アイテム、またはタスクが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ExternalTaskCommentListResponse": {
        "required": [
          "workspace",
          "item",
          "task",
          "comments"
        ],
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/ExternalWorkspaceRefResponse"
          },
          "item": {
            "$ref": "#/components/schemas/ExternalItemRefResponse"
          },
          "task": {
            "$ref": "#/components/schemas/ExternalTaskRefResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalTaskCommentResponse"
            },
            "description": "コメント一覧"
          }
        },
        "description": "外部公開用タスクコメント一覧レスポンス"
      },
      "ExternalTaskCommentResponse": {
        "required": [
          "id",
          "user",
          "content"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "コメントID",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/ExternalUserRefResponse"
          },
          "content": {
            "type": "string",
            "description": "コメント内容"
          },
          "commentType": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/TaskCommentType"
              }
            ]
          }
        },
        "description": "外部公開用タスクコメント"
      },
      "TaskCommentType": {
        "enum": [
          "Normal",
          "Memo",
          "HelpWanted",
          "NeedReply",
          "Reminder",
          "Urge"
        ],
        "type": "string",
        "nullable": true
      },
      "ExternalUserRefResponse": {
        "required": [
          "loginId",
          "username"
        ],
        "type": "object",
        "properties": {
          "loginId": {
            "type": "string",
            "description": "ログインID"
          },
          "username": {
            "type": "string",
            "description": "ユーザー名"
          }
        },
        "description": "外部公開用ユーザー参照"
      },
      "ExternalTaskRefResponse": {
        "required": [
          "sequence"
        ],
        "type": "object",
        "properties": {
          "sequence": {
            "type": "integer",
            "description": "シーケンス番号",
            "format": "int32"
          }
        },
        "description": "外部公開用タスク参照"
      },
      "ExternalItemRefResponse": {
        "required": [
          "itemNumber"
        ],
        "type": "object",
        "properties": {
          "itemNumber": {
            "type": "integer",
            "description": "アイテム番号（ワークスペース内連番）",
            "format": "int32"
          }
        },
        "description": "外部公開用アイテム参照"
      },
      "ExternalWorkspaceRefResponse": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "ワークスペースコード"
          }
        },
        "description": "外部公開用ワークスペース参照"
      },
      "ExternalWorkspaceTaskDetailResponse": {
        "required": [
          "workspace",
          "item",
          "task"
        ],
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/ExternalWorkspaceRefResponse"
          },
          "item": {
            "$ref": "#/components/schemas/ExternalItemRefResponse"
          },
          "task": {
            "$ref": "#/components/schemas/ExternalTaskDetailResponse"
          }
        },
        "description": "外部公開用タスク詳細レスポンス"
      },
...
      "ExternalTaskDiscardedResponse": {
        "required": [
          "discardedAt"
        ],
        "type": "object",
        "properties": {
          "discardedAt": {
            "type": "string",
            "description": "破棄日時",
            "format": "date-time",
            "nullable": true
          },
          "discardReason": {
            "type": "string",
            "description": "破棄理由",
            "nullable": true
          }
        },
        "description": "外部公開用タスク破棄情報"
      },
      "TaskPriority": {
        "enum": [
          "Low",
          "Medium",
          "High",
          "Critical"
        ],
        "type": "string",
        "nullable": true
      },
      "ExternalTaskTypeRefResponse": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "タスク種類コード"
          },
          "name": {
            "type": "string",
            "description": "タスク種類名"
          }
        },
        "description": "外部公開用タスク種類参照"
      },
      "ExternalWorkspaceTaskListResponse": {
        "required": [
          "workspace",
          "item",
          "tasks"
        ],
        "type": "object",
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/ExternalWorkspaceRefResponse"
          },
          "item": {
            "$ref": "#/components/schemas/ExternalItemRefResponse"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalTaskSummaryResponse"
            },
            "description": "タスク一覧"
          }
        },
        "description": "外部公開用タスク一覧レスポンス"
      },
      "ExternalTaskSummaryResponse": {
        "required": [
          "sequence",
          "assignedUser",
          "createdByUser",
          "isCompleted",
          "taskType",
          "dueDate",
          "progressPercentage"
        ],
        "type": "object",
        "properties": {
          "sequence": {
            "type": "integer",
            "description": "シーケンス番号（アイテム内一意）",
            "format": "int32"
          },
          "assignedUser": {
            "$ref": "#/components/schemas/ExternalUserRefResponse"
          },
          "createdByUser": {
            "$ref": "#/components/schemas/ExternalUserRefResponse"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "完了フラグ"
          },
          "completedAt": {
            "type": "string",
            "description": "完了日時",
            "format": "date-time",
            "nullable": true
          },
          "completedByUser": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalUserRefResponse"
              }
            ]
          },
          "discarded": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalTaskDiscardedResponse"
              }
            ]
          },
          "taskType": {
            "$ref": "#/components/schemas/ExternalTaskTypeRefResponse"
          },
          "priority": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ]
          },
          "startDate": {
            "type": "string",
            "description": "開始日時",
            "format": "date-time",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "description": "期限日時",
            "format": "date-time"
          },
          "estimatedHours": {
            "type": "number",
            "description": "予定工数（時間）",
            "format": "double",
            "nullable": true
          },
          "actualHours": {
            "type": "number",
            "description": "実績工数（時間）",
            "format": "double",
            "nullable": true
          },
          "progressPercentage": {
            "type": "integer",
            "description": "進捗率（0-100）",
            "format": "int32"
          }
        },
        "description": "外部公開用タスクサマリー（一覧用）"
      },
      "ExternalItemResponse": {
        "required": [
          "workspaceCode",
          "itemNumber",
          "subject",
          "tags",
          "owner",
          "isActive",
          "isArchived",
          "isDraft"
        ],
        "type": "object",
        "properties": {
          "workspaceCode": {
            "type": "string",
            "description": "ワークスペースコード"
          },
          "itemNumber": {
            "type": "integer",
            "description": "アイテム番号（ワークスペース内の連番）",
            "format": "int32"
          },
          "subject": {
            "type": "string",
            "description": "件名"
          },
          "body": {
            "type": "string",
            "description": "本文（Markdown変換データ）",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "タグ（タグ名一覧）"
          },
          "owner": {
            "$ref": "#/components/schemas/ExternalUserRefResponse"
          },
          "assignedUser": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalUserRefResponse"
              }
            ]
          },
          "committer": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalUserRefResponse"
              }
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "期限日時",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブフラグ"
          },
          "isArchived": {
            "type": "boolean",
            "description": "アーカイブフラグ"
          },
          "isDraft": {
            "type": "boolean",
            "description": "下書きフラグ"
          }
        },
        "description": "外部公開用アイテム情報レスポンス"
      },
      "ExternalItemListResponse": {
        "required": [
          "workspaceCode",
          "totalCount",
          "currentPage",
          "pageSize",
          "hasNextPage",
          "items"
        ],
        "type": "object",
        "properties": {
          "workspaceCode": {
            "type": "string",
            "description": "ワークスペースコード"
          },
          "totalCount": {
            "type": "integer",
            "description": "総件数",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "現在のページ",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "ページあたりの件数",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "次ページの有無"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalItemSummaryResponse"
            },
            "description": "アイテムの配列"
          }
        },
        "description": "外部公開用アイテム一覧レスポンス"
      },
      "ExternalItemSummaryResponse": {
        "required": [
          "itemNumber",
          "subject",
          "tags",
          "owner",
          "isActive",
          "isArchived",
          "isDraft"
        ],
        "type": "object",
        "properties": {
          "itemNumber": {
            "type": "integer",
            "description": "アイテム番号（ワークスペース内の連番）",
            "format": "int32"
          },
          "subject": {
            "type": "string",
            "description": "件名"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "タグ（タグ名一覧）"
          },
          "owner": {
            "$ref": "#/components/schemas/ExternalUserRefResponse"
          },
          "assignedUser": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalUserRefResponse"
              }
            ]
          },
          "committer": {
            "oneOf": [
              {
                "enum": [
                  null
                ],
                "nullable": true
              },
              {
                "$ref": "#/components/schemas/ExternalUserRefResponse"
              }
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "期限日時",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "アクティブフラグ"
          },
          "isArchived": {
            "type": "boolean",
            "description": "アーカイブフラグ"
          },
          "isDraft": {
            "type": "boolean",
            "description": "下書きフラグ"
          }
        },
        "description": "外部公開用アイテムサマリーレスポンス（一覧用）"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "External API key sent in the X-API-KEY header.",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  },
...
}